🚚 move #3: 컴포넌트 폴더 구조 수정
- common 및 page 단위로 정리
