+++ src/components/home/Logo/LogoCharacter/index.tsx
