+++ src/components/size/SizeFooter/index.tsx
@@ -5,21 +5,20 @@ import { useUserSelectStore } from '@/stores/useUserSelectStore';
 
 type Props = {
   nextPage: string;
-  text?: string;
 };
 
-export default function FooterButton({ nextPage, text = '다음' }: Props) {
+export default function SizeFooter({ nextPage }: Props) {
   const userSelect = useUserSelectStore(state => state.userSelect);
-  const isSelected = userSelect.sizeId !== 0;
+  const disabled = userSelect.sizeId === 0;
 
   return (
     <footer className="flex h-20 w-full flex-row justify-around border-t border-t-gray_00-light p-2 pb-8 dark:border-t-gray_00-dark">
       <Link
-        href={isSelected ? nextPage : '#'}
+        href={disabled ? '#' : nextPage}
         className={`flex h-12 w-[340px] items-center justify-center rounded-lg border-0 bg-primary-light px-6 text-medium font-semi_bold text-white-light
-        ${isSelected || 'cursor-not-allowed opacity-50'}`}
+        ${disabled && 'cursor-not-allowed opacity-50'}`}
       >
-        {text}
+        다음
       </Link>
     </footer>
   );
